requirements.txt
deta==1.2.0
plotly==5.18.0
python-dotenv==1.0.0
streamlit==1.28.2
streamlit_option_menu==0.3.6
